+++ examples/data-sources/verumex_warehouse_metadata/data-source.tf
@@ -1,0 +1,3 @@
+data "verumex_warehouse_metadata" "example" {
+  warehouse_id = "foo"
+}