+++ examples/resources/verumex_company/resource.tf
@@ -1,0 +1,3 @@
+resource "verumex_company" "system_admin" {
+  name = "System Administrators"
+}