+++ internal/provider/provider.go
@@ -27,10 +27,10 @@ func New(version string) func() *schema.Provider {
 	return func() *schema.Provider {
 		p := &schema.Provider{
 			DataSourcesMap: map[string]*schema.Resource{
-				"scaffolding_data_source": dataSourceScaffolding(),
+				"verumex_warehouse_metadata": verumexWarehouseMetadata(),
 			},
 			ResourcesMap: map[string]*schema.Resource{
-				"scaffolding_resource": resourceScaffolding(),
+				"verumex_company": verumexCompany(),
 			},
 		}
 
@@ -48,9 +48,9 @@ type apiClient struct {
 
 func configure(version string, p *schema.Provider) func(context.Context, *schema.ResourceData) (any, diag.Diagnostics) {
 	return func(context.Context, *schema.ResourceData) (any, diag.Diagnostics) {
-		// Setup a User-Agent for your API client (replace the provider name for yours):
-		// userAgent := p.UserAgent("terraform-provider-scaffolding", version)
-		// TODO: myClient.UserAgent = userAgent
+		// Setup a User-Agent for your API client:
+		// userAgent := p.UserAgent("verumex", version)
+		// myClient.UserAgent = userAgent
 
 		return &apiClient{}, nil
 	}