+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.1.0 (Unreleased)
+## 0.1.0
 
-BACKWARDS INCOMPATIBILITIES / NOTES:
+NOTES:
+
+This is an initial version that bootstraps the provider with sample resource
+definitions based on the Terraform Provider template
+[scaffold](https://github.com/hashicorp/terraform-provider-scaffolding).